refactor(search): fold property index reader into properties_db_client
Move get_entity_properties_for_index and IndexedEntityProperty into
properties_db_client (the existing property-DB-utils crate) and delete the
redundant entity_property_db_utils crate. Replace the sub_type string match
in properties_entity_type with a From<DocumentSubType> for EntityType
conversion in models_properties, reused by the search_service indexer site.
