--- rust/cloud-storage/entity_property_db_utils/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-edition = "2024"
-name = "entity_property_db_utils"
-publish = false
-version = "0.1.0"
-
-[dependencies]
-anyhow = { workspace = true }
-models_properties = { path = "../models_properties" }
-serde_json = { workspace = true }
-sqlx = { workspace = true }
-tracing = { workspace = true }
-uuid = { workspace = true }
-workspace-hack = { version = "0.1", path = "../workspace-hack" }
