+++ rust/cloud-storage/models_properties/Cargo.toml
@@ -6,6 +6,7 @@ version = "0.1.0"
 
 [dependencies]
 chrono = { workspace = true }
+document_sub_type = { path = "../document_sub_type" }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_with = { workspace = true }
