+++ rust/cloud-storage/search_processing_service/Cargo.toml
@@ -8,7 +8,7 @@ version = "0.1.0"
 [features]
 default = ["processing", "service"]
 processing = []
-service = ["macro_db_client", "models_properties", "entity_property_db_utils"]
+service = ["macro_db_client", "models_properties", "properties_db_client"]
 
 [[bin]]
 name = "search_processing_service"
@@ -25,7 +25,6 @@ axum = { workspace = true }
 chrono = { workspace = true }
 comms_db_client = { path = "../comms_db_client" }
 email_db_client = { path = "../email_db_client" }
-entity_property_db_utils = { path = "../entity_property_db_utils", optional = true }
 futures = { workspace = true }
 http-body-util = { workspace = true }
 lexical_client = { path = "../lexical_client" }
@@ -53,6 +52,7 @@ nom = { workspace = true }
 opensearch_client = { path = "../opensearch_client" }
 ensure_exists = { path = "../ensure_exists", features = ["dynamodb"] }
 pdfium-render = { workspace = true }
+properties_db_client = { path = "../properties_db_client", optional = true }
 rust-embed = "8.6.0"
 s3_client = { path = "../s3_client" }
 secretsmanager_client = { path = "../secretsmanager_client" }
